+++ CHANGELOG.md
@@ -5,6 +5,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+## Added
+- Add "immediately" parameter to "cancel" method in BillingAgreement
+- Add "cancel_immediately" and "expire" actions to BillingAgreementAdmin
+
 ## Fixed
 - Fix "state" field having values of both "Canceled" and "Cancelled" in BillingAgreement model